refactor: move d1 to vectors
