+++ src/coordinax/_src/compat.py
@@ -12,12 +12,16 @@ import quaxed.numpy as xp
 from dataclassish import field_values
 from unxt.quantity import AbstractQuantity, Quantity, UncheckedQuantity
 
-from coordinax._src.d1.cartesian import CartesianAcc1D, CartesianVel1D
-from coordinax._src.d1.radial import RadialAcc, RadialVel
 from coordinax._src.d2.cartesian import CartesianAcc2D, CartesianVel2D
 from coordinax._src.d3.cartesian import CartesianAcc3D, CartesianVel3D
 from coordinax._src.utils import full_shaped
 from coordinax._src.vectors.base.base import AbstractVector
+from coordinax._src.vectors.d1 import (
+    CartesianAcc1D,
+    CartesianVel1D,
+    RadialAcc,
+    RadialVel,
+)
 
 #####################################################################
 # Convert to Quantity
