--- src/coordinax/_src/d1/__init__.py
@@ -1,16 +1,0 @@
-# pylint: disable=duplicate-code
-"""1-dimensional."""
-
-from . import base, cartesian, compat, radial, transform
-from .base import *
-from .cartesian import *
-from .compat import *
-from .radial import *
-from .transform import *
-
-__all__: list[str] = []
-__all__ += base.__all__
-__all__ += cartesian.__all__
-__all__ += radial.__all__
-__all__ += transform.__all__
-__all__ += compat.__all__
