+++ src/coordinax/_src/transform/d2.py
@@ -10,8 +10,6 @@ from plum import dispatch
 import quaxed.numpy as jnp
 from unxt.quantity import AbstractQuantity, Quantity
 
-from coordinax._src.d1.cartesian import CartesianPos1D
-from coordinax._src.d1.radial import RadialPos
 from coordinax._src.d2.base import AbstractPos2D
 from coordinax._src.d2.cartesian import CartesianPos2D
 from coordinax._src.d2.polar import PolarPos
@@ -21,6 +19,7 @@ from coordinax._src.d3.cylindrical import CylindricalPos
 from coordinax._src.d3.mathspherical import MathSphericalPos
 from coordinax._src.d3.spherical import SphericalPos
 from coordinax._src.exceptions import IrreversibleDimensionChange
+from coordinax._src.vectors.d1 import CartesianPos1D, RadialPos
 
 
 @dispatch.multi(
