+++ src/coordinax/_src/transform/differentials.py
@@ -14,11 +14,11 @@ import quaxed.numpy as jnp
 from dataclassish import field_items
 from unxt import Quantity
 
-from coordinax._src.d1.base import AbstractVel1D
 from coordinax._src.d2.base import AbstractVel2D
 from coordinax._src.d3.base import AbstractVel3D
 from coordinax._src.distance.base import AbstractDistance
 from coordinax._src.vectors.base import AbstractPos, AbstractVel
+from coordinax._src.vectors.d1 import AbstractVel1D
 
 
 # TODO: implement for cross-representations
