+++ src/coordinax/_src/vectors/d1/base.py
