+++ src/coordinax/_src/vectors/d1/cartesian.py
