+++ src/coordinax/_src/vectors/d1/compat.py
